requirements.txt
telethon==1.28.5
python-telegram-bot==20.7
sqlalchemy==2.0.23
schedule==1.2.0
requests==2.32.3
aiohttp==3.9.5
flask==3.0.3
python-dotenv==1.0.1
httpx==0.25.2